+++ rbt_classes.cpp
@@ -329,14 +329,14 @@ void InformationSet::get_actions_given_policy(std::vector<int>& actions, PolicyV
     else {
         // std::cout << "Index: " << this->get_index() << " for hash: " << this->get_hash() << std::endl;
         if (this->move_flag) {
-            std::vector<float>& prob_dist = policy_obj.policy_dict[this->get_index()];
+            std::vector<double>& prob_dist = policy_obj.policy_dict[this->get_index()];
             for (int move = 0; move < 9; move++) {
                 if (prob_dist[move] > 0) {
                     actions.push_back(move);
                 }
             }
         } else {
-            std::vector<float>& prob_dist = policy_obj.policy_dict[this->get_index()];
+            std::vector<double>& prob_dist = policy_obj.policy_dict[this->get_index()];
             for (int sense = 9; sense < 13; sense++) {
                 if (prob_dist[sense] > 0) {
                     actions.push_back(sense);
@@ -619,7 +619,7 @@ void History::get_information_sets(InformationSet &I_1, InformationSet &I_2) {
     }
 }
 
-TerminalHistory::TerminalHistory(std::vector<int>& history, std::vector<float> reward) : History(history) {
+TerminalHistory::TerminalHistory(std::vector<int>& history, std::vector<double> reward) : History(history) {
     if (reward.empty()) {
         this->reward = {0.0, 0.0};
     } else {
@@ -671,7 +671,7 @@ NonTerminalHistory NonTerminalHistory::copy() {
 
 Policy::Policy() {
     this->player = '0';
-    this->policy_dict = std::unordered_map<std::string, std::vector<float> >();
+    this->policy_dict = std::unordered_map<std::string, std::vector<double> >();
 }
 
 Policy::Policy(char player, std::string& file_path) {
@@ -679,7 +679,7 @@ Policy::Policy(char player, std::string& file_path) {
     this->policy_dict = this->read_policy_from_json(file_path, player);
 }
 
-Policy::Policy(char player, std::unordered_map<std::string, std::vector<float> >& policy_dict) {
+Policy::Policy(char player, std::unordered_map<std::string, std::vector<double> >& policy_dict) {
     this->player = player;
     this->policy_dict = policy_dict;
 }
@@ -688,22 +688,22 @@ Policy Policy::copy() {
     return Policy(this->player, this->policy_dict);
 }
 
-void Policy::update_policy_for_given_information_set(InformationSet& information_set, std::vector<float>& prob_distribution) {
-    std::vector<float> prob_dist;
+void Policy::update_policy_for_given_information_set(InformationSet& information_set, std::vector<double>& prob_distribution) {
+    std::vector<double> prob_dist;
     for (int i = 0; i < prob_distribution.size(); i++) {
         prob_dist[i] = prob_distribution[i];
     }
     this->policy_dict[information_set.get_hash()] = prob_dist;
 }
 
-std::unordered_map<std::string, std::vector<float> > Policy::read_policy_from_json(std::string& file_path, char player){
+std::unordered_map<std::string, std::vector<double> > Policy::read_policy_from_json(std::string& file_path, char player){
     std::ifstream i(file_path);
     json policy_obj;
     i >> policy_obj;
     
     for (json::iterator it = policy_obj.begin(); it != policy_obj.end(); ++it) {
         std::string key = it.key();
-        std::vector <float> probability_distribution(13);
+        std::vector <double> probability_distribution(13);
         // initialise all values to zero
         for (int i = 0; i < 13; i++) {
             probability_distribution[i] = 0;
@@ -745,7 +745,7 @@ std::unordered_map<std::string, std::vector<float> > Policy::read_policy_from_js
 
 PolicyVec::PolicyVec() {
     this->player = '0';
-    this->policy_dict = std::vector< std::vector<float> >();
+    this->policy_dict = std::vector< std::vector<double> >();
 }
 
 PolicyVec::PolicyVec(char player, std::string& file_path) {
@@ -753,7 +753,7 @@ PolicyVec::PolicyVec(char player, std::string& file_path) {
     this->policy_dict = this->read_policy_from_json(file_path, player);
 }
 
-PolicyVec::PolicyVec(char player, std::vector< std::vector<float> >& policy_dict) {
+PolicyVec::PolicyVec(char player, std::vector< std::vector<double> >& policy_dict) {
     this->player = player;
     this->policy_dict = policy_dict;
 }
@@ -762,17 +762,17 @@ PolicyVec PolicyVec::copy() {
     return PolicyVec(this->player, this->policy_dict);
 }
 
-void PolicyVec::update_policy_for_given_information_set(InformationSet& I, std::vector<float>& prob_distribution) {
-    std::vector<float> prob_dist;
+void PolicyVec::update_policy_for_given_information_set(InformationSet& I, std::vector<double>& prob_distribution) {
+    std::vector<double> prob_dist;
     for (int i = 0; i < prob_distribution.size(); i++) {
         prob_dist[i] = prob_distribution[i];
     }
     this->policy_dict[I.get_index()] = prob_dist;
 }
 
-std::vector< std::vector<float> > PolicyVec::read_policy_from_json(std::string& file_path, char player){
+std::vector< std::vector<double> > PolicyVec::read_policy_from_json(std::string& file_path, char player){
     long int policy_size = player == 'x' ? InformationSet::P1_hash_to_int_map.size() : InformationSet::P2_hash_to_int_map.size();
-    std::vector< std::vector<float> > policy_list(policy_size);
+    std::vector< std::vector<double> > policy_list(policy_size);
     
     std::ifstream i(file_path);
     json policy_obj;
@@ -790,7 +790,7 @@ std::vector< std::vector<float> > PolicyVec::read_policy_from_json(std::string& 
 
         InformationSet I(player, move_flag, I_hash);
 
-        std::vector <float> probability_distribution(13);
+        std::vector <double> probability_distribution(13);
         // initialise all values to zero
         for (int i = 0; i < 13; i++) {
             probability_distribution[i] = 0;
